[DEVELOPMENT ENVIRONMENT] * BUG: Don't check conflict on project configuration file if that files   does not exist.
[INTERPRETER]
* OPT: Move the string splitting routine in its own file and optimize it.

[INFORMER]
* BUG: Remove unneeded parenthesis on a test.

[GB.CLIPPER]
* BUG: Remove an unused variable.

[GB.DESKTOP]
* BUG: Fix error message in DesktopMime.

[GB.FORM]
* BUG: MaskBox: Raise the Change event after moving the cursor, so that it 
  behaves exactly like a TextBox.

[GB.FORM.MDI]
* NEW: Workspace: Count the number of detached windows so that the "detach 
  current tab" menu entry is disabled when needed.

[GB.GTK]
* BUG: Window.Utility should work even if the window is mapped now.

[GB.GTK3]
* BUG: Window.Utility should work even if the window is mapped now.

[GB.NET.PO3]
* NEW: Add a void constructor to the Pop3Client class for the 
  documentation.


git-svn-id: svn://localhost/gambas/trunk@6285 867c0c6c-44f3-4631-809d-bfa615b0a4ec

+++ main/share/gb_code_temp.h
@@ -205,7 +205,7 @@ void CODE_end_function(FUNCTION *func)
 
 static ushort *get_last_code()
 {
-	if (cur_func->last_code == CODE_NO_POS)
+	if ((ushort)cur_func->last_code == CODE_NO_POS)
 		return NULL;
 
 	return &cur_func->code[cur_func->last_code];
@@ -213,7 +213,7 @@ static ushort *get_last_code()
 
 static ushort *get_last_code2()
 {
-	if (cur_func->last_code2 == CODE_NO_POS)
+	if ((ushort)cur_func->last_code2 == CODE_NO_POS)
 		return NULL;
 
 	return &cur_func->code[cur_func->last_code2];
